fix(crop): keep float box semantics in Rust core
First divergence: Pillow Image.crop accepts numeric boxes, rounds coordinates with ties-to-even, applies the decompression-bomb extent check, and then enforces signed C-long coordinate limits; the Rust wrapper previously extracted only i32 coordinates and rejected floats or large offsets before core logic.\n\nMove rounding, NaN/infinity/finite overflow handling, extent protection, and signed-range errors into the Rust crop core while keeping PyO3 and the Python facade as thin marshaling/delegation layers. Add input-only parity cases for fractional and boundary numeric boxes.\n\nLane impact: 11 crop cases added; canonical parity is 2714/2715 with the sole failure remaining in the separately owned fontdone variable-font lane. Image-core region evidence increased from 13238/14077 (94.0%) to 13339/14182 (94.1%).

+++ pillow-rs-py/python/pillow_rs/image.py
@@ -184,7 +184,7 @@ class Image:
         del reducing_gap
         return Image(self._rust_image.resize(size, resample, box))
 
-    def crop(self, box: Optional[Tuple[int, int, int, int]] = None) -> "Image":
+    def crop(self, box: Optional[Tuple[float, float, float, float]] = None) -> "Image":
         return Image(self._rust_image.crop(box))
 
     def rotate(
